Always use maq 0.6.3 to do submap and mapmerge.
In AcceptReads/Maq, don't remove the whole-lane map files for Genome Models
with newer maq

git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk@33945 127c0c0e-971a-0410-ad48-868c1a942d8d

+++ Genome/Model/Command/AddReads/AcceptReads/Maq.pm
@@ -42,7 +42,6 @@ sub execute {
 
     $DB::single = 1;
     
-    # ensure the reference sequence exists.
     my $lane = $self->run->limit_regions;
     unless ($lane) {
         $self->error_message("No limit regions parameter on run_id ".$self->run_id);
@@ -76,7 +75,8 @@ sub execute {
             }
         }
 
-        system($maq_pathname, 'mapmerge', $lane_mapfile, @input_mapfiles);
+        #system($maq_pathname, 'mapmerge', $lane_mapfile, @input_mapfiles);
+        system('/gsc/pkg/bio/maq/maq-0.6.3_x86_64-linux/maq', 'mapmerge', $lane_mapfile, @input_mapfiles);
     }
         
     unless (-f $lane_mapfile) {
@@ -88,9 +88,19 @@ sub execute {
     my ($evenness)=($line=~/(\S+)\%$/);
     if($evenness > $model->align_dist_threshold) {
         # The align-reads step make submap files for each chromosome.  We can delete this one now
-        unlink $lane_mapfile;
-        unlink $self->unique_alignment_file_for_lane;
-        unlink $self->duplicate_alignment_file_for_lane;
+        if ($model->read_aligner_name ne 'maq0_6_3') {
+            # FIXME For 0.6.4 and 0.6.5, don't remove the whole-lane map files, only any new file
+            # we may have created in /tmp.  When we're convinced that we can submap and mapmerge
+            # successfully with newer maq's, the, we only need to keep the three unlink()s
+            # in the else block below
+            if ($lane_mapfile =~ m#/tmp/AcceptReads#) {
+                unlink $lane_mapfile;
+            }
+        } else {
+            unlink $lane_mapfile;
+            unlink $self->unique_alignment_file_for_lane;
+            unlink $self->duplicate_alignment_file_for_lane;
+        }
 
         return 1;
     } else {
